refactor swaps liveness reducer, set stx status at the same time
